% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.R
\name{ipeds_context}
\alias{ipeds_context}
\title{Print IPEDS context for LLM integration}
\usage{
ipeds_context()
}
\value{
Invisibly returns the context string
}
\description{
Print IPEDS context for LLM integration
}
