% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.R
\name{ipeds_enrollment}
\alias{ipeds_enrollment}
\title{Search IPEDS fall enrollment data}
\usage{
ipeds_enrollment(year = 2021, limit = 100)
}
\arguments{
\item{year}{Academic year (e.g., 2021). Default 2021.}

\item{limit}{Maximum records to return. Default 100.}
}
\value{
tibble: unitid, inst_name, year, enrollment_fall, state_abbr
}
\description{
Returns enrollment data from IPEDS.
}
